requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.6.2
geventhttpclient==1.4.4
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
locust==1.1.1
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.19.0
pandas==1.0.5
plotly==4.8.2
psutil==5.7.0
pur==5.3.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
yattag==1.13.2
zope.event==4.4
zope.interface==5.1.0
